Rename memory header file
