+++ .vscode/c_cpp_properties.json
@@ -7,7 +7,7 @@
                 "${workspaceFolder}/temp"
             ],
             "defines": [],
-            "compilerPath": "/usr/bin/clang"
+            "compilerPath": "/usr/bin/gcc"
         }
     ],
     "version": 4
