+++ GB_Emulator_C.code-workspace
@@ -11,7 +11,11 @@
 			"stdio.h": "c",
 			"errno.h": "c",
 			"errno-base.h": "c",
-			"features.h": "c"
+			"features.h": "c",
+			"libc-header-start.h": "c",
+			"stubs.h": "c",
+			"registers.h": "c",
+			"endian.h": "c"
 		},
 		"C_Cpp.default.cppStandard": "c++17",
 		"C_Cpp.default.cStandard": "c17",
