+++ include/registers.h
@@ -2,8 +2,13 @@ struct registers {
     struct {
         union {
             struct {
+                #ifdef LITTLE_ENDIAN
+                unsigned char F;
+                unsigned char A;
+                #else
                 unsigned char A;
                 unsigned char F;
+                #endif
             };
             unsigned short AF;
         };
@@ -11,8 +16,13 @@ struct registers {
     struct {
         union {
             struct {
+                #ifdef LITTLE_ENDIAN
+                unsigned char C;
+                unsigned char B;
+                #else
                 unsigned char B;
                 unsigned char C;
+                #endif
             };
             unsigned short BC;
         };
@@ -20,8 +30,13 @@ struct registers {
     struct {
         union {
             struct {
+                #ifdef LITTLE_ENDIAN
+                unsigned char E;
+                unsigned char D;
+                #else
                 unsigned char D;
                 unsigned char E;
+                #endif
             };
             unsigned short DE;
         };
@@ -29,8 +44,13 @@ struct registers {
     struct {
         union {
             struct {
+                #ifdef LITTLE_ENDIAN
+                unsigned char L;
+                unsigned char H;
+                #else
                 unsigned char H;
                 unsigned char L;
+                #endif
             };
             unsigned short HL;
         };
@@ -38,12 +58,3 @@ struct registers {
     unsigned short SP;
     unsigned short PC;
 } extern registers;
-
-enum eRegisters {
-    A, F, AF,
-    B, C, BC,
-    D, E, DE,
-    H, L, HL,
-    SP, PC,
-    numRegisters
-};
